work on chat input and server movement foundation
enter sends chat. escape moves focus back to gameworld. arrow keys send
direction specific movement commands to server. client will have only
local x and y positioning for drawing things onto the screen correctly,
while server will maintain the x y and z of objects and people for
tracking and logic.
also came up with a way for server to send things between clients. ip
address will be set on login for clients as an account value. the
accounts will also have a String[] that contains info that needs to be
sent to that client. on the write loop, if that client has things in the
string[], it will write those to the client, then exit the write loop. i
think. not sure if it will work well.