Receive surrounding players on login && more
Fixed a few account creation and login bugs. Optimized code that updates
accounts map on logout and disconnect.  Added some colors to tile
generator for easier use in the future. Made code easier to read
everywhere. Put removeClosedAccounts method back in the server class. It
has nothing to do with packets, so doesn't belong in packethandler.