Added system status panel and server optimization
Added a panel which displays fps, and will display ping in the future.
Also removed notifySelfAllInRange() because it did the exact same thing
as notifyAllInRange. Was senseless for it to do two loops, when it could
simply do 1, and send to either current client, or all clients online
based on a switch. So, that's what it does now.